+++ src/main/java/seedu/address/model/Model.java
@@ -85,6 +85,12 @@ public interface Model {
     /** Returns an unmodifiable view of the filtered vendor list */
     ObservableList<Person> getFilteredVendorList();
 
+    /** Returns the count of the filtered guest list in total and by rsvp status */
+    int[] getFilteredGuestListCount();
+
+    /** Returns the total count of the filtered vendor list */
+    int getFilteredVendorListCount();
+
     /**
      * Updates the filter of the filtered person list to filter by the given {@code predicate}.
      * @throws NullPointerException if {@code predicate} is null.
